Added a switch for the styles. Moved the log-management related stories to a folder
